Needs to be looked at by Auth expert or Auth0 support. My final commit with the package added.
When bringing in Auth0, it goes to a Login.aspx page, which we don't have in the proj. Tutorial at https://auth0.com/docs/quickstart/webapp/aspnet was followed. Auth0 Support needed at this point, but they were unavailable during the weekend. 'Was going to submit ticket but bot did not respond to it.